backend/simple_services/post/requirements.txt
asyncmy==0.27
fastapi==0.89.1
mysql-connector-python==8.0.32
pydantic==1.10.4
python-dotenv==0.21.0
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.39.0
strawberry-graphql==0.159.0
uvicorn==0.20.0

